refactor campaign grid into component
